# backend/src/main/resources/openapi/api.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "HEH Gourmet",
        "description": "API for the HEH cafeteria online project",
        "version": "0.0.1",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "contact": {
            "name": "HEH Gourmet Support"
        }
    },
    "servers": [
        {
            "description": "localhost dev server",
            "url": "https:localhost:8080"
        },
        {
            "description": "Production server",
            "url": "https://gourmet-iso4.onrender.com"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "OAuth": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {
                            "write": "Grants write access"
                        }
                    }
                }
            }
        },
...
        "schemas": {
...
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "float"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "image": {
                        "type": "url"
                    },
                    "categoryID": {
                        "type": "integer"
                    }
                }
            },
            "InputProduct": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "float"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "image": {
                        "type": "url"
                    },
                    "categoryID": {
                        "type": "integer"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    }
                }
            },
            "InputUser": {
                "type": "object",
                "properties": {
                    "lastname": {
                        "type": "string"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get all products",
                "operationId": "getProducts",
                "requestBody": {
                    "description": "Category to get products from",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "integer"
                            }
                        }
                    },
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "A list of product.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Batch add of product",
                "requestBody": {
                    "description": "Products to add",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Product created"
                    }
                }
            },
            "delete": {
                "summary": "Batch delete of product",
                "responses": {
                    "204": {
                        "description": "Products deleted"
                    }
                }
            }
        },
        "/api/products/{category_id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get all product from a category",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All product from a category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Add a product",
                "requestBody": {
                    "description": "Product to add",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InputProduct"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Product created"
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get a product by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A product.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Update a product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Product to update",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InputProduct"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Product updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete a product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted"
                    }
                }
            }
        },
        "/api/user/login": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user data from email",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get User Data including ID , last name , first name , etc",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register new user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InputUser"
                            }
                        }
                    }
                }
            }
        },
        "/api/user/promote/{id}": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Promote user from customer to another Role ex: Admin , Employee, etc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "ADMIN",
                                            "EMPLOYEE",
                                            "CUSTOMER"
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User promoted"
                    }
                }
            }
        },
...
    }
}